+++ UnitTestFundamental/Credential.swift
@@ -1,0 +1,15 @@
+//
+//  Credential.swift
+//  UnitTestFundamental
+//
+//  Created by Shad Mazumder on 28/10/23.
+//
+
+import Foundation
+
+public struct Credential {
+    public let username: String
+    public let password: String
+}
+
+extension Credential: Encodable{}
